@@ -1,68 +1,83 @@
 {
-	"name": "everdev",
-	"version": "1.1.2",
-	"description": "Everscale Dev Environment",
-	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/tonlabs/everdev.git"
-	},
-	"scripts": {
-		"prepublishOnly": "npm i && tsc",
-		"everdev": "node ./cli.js",
-		"test": "jest -i"
-	},
-	"bin": {
-		"everdev": "./cli.js"
-	},
-	"keywords": [
-		"Free",
-		"TON",
-		"Ever",
-		"EverScale",
-		"Freeton",
-		"Dev",
-		"Development",
-		"TONOS",
-		"EVEROS"
-	],
-	"author": "TON DEV SOLUTIONS LTD.",
-	"license": "Apache-2.0",
-	"licenses": [
-		{
-			"type": "Apache-2.0",
-			"url": "http://www.apache.org/licenses/LICENSE-2.0"
-		}
-	],
-	"homepage": "https://docs.ton.dev/",
-	"engines": {
-		"node": ">=6"
-	},
-	"dependencies": {
-		"@tonclient/appkit": "^0",
-		"@tonclient/core": "^1",
-		"@tonclient/lib-node": "^1",
-		"chalk": "^2.4.2",
-		"dockerode": "^3.3.1",
-		"fs-extra": "^9.1.0",
-		"mv": "^2.1.1",
-		"request": "^2.88.2",
-		"unzip-stream": "^0.3.1"
-	},
-	"devDependencies": {
-		"@babel/preset-env": "^7.14.7",
-		"@babel/preset-typescript": "^7.14.5",
-		"@types/dockerode": "^3.2.2",
-		"@types/fs-extra": "^9.0.8",
-		"@types/jest": "^27.1.4",
-		"@types/mv": "^2.1.1",
-		"@types/node": "^12.12.6",
-		"@types/request": "^2.48.5",
-		"@types/unzip-stream": "^0.3.0",
-		"caxa": "^1.0.0",
-		"jest": "^27.1.4",
-		"ts-jest": "^27.1.4",
-		"typescript": "^4.1.3"
-	}
+    "name": "everdev",
+    "version": "1.1.3",
+    "description": "Everscale Dev Environment",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/tonlabs/everdev.git"
+    },
+    "scripts": {
+        "prepublishOnly": "npm i && tsc",
+        "everdev": "node ./cli.js",
+        "format": "prettier --write 'src/**/*.ts' 'test/**/*.ts'",
+        "eslint": "eslint \"**/*.ts\"",
+        "eslint-fix": "eslint \"**/*.ts\" --fix",
+        "tsc": "tsc",
+        "build": "rimraf dist && npm run eslint &&  tsc --project tsconfig.build.json",
+        "test": "jest -i --forceExit --detectOpenHandles --coverage --verbose"
+    },
+    "bin": {
+        "everdev": "./cli.js"
+    },
+    "keywords": [
+        "Free",
+        "TON",
+        "Ever",
+        "EverScale",
+        "Freeton",
+        "Dev",
+        "Development",
+        "TONOS",
+        "EVEROS"
+    ],
+    "author": "TON DEV SOLUTIONS LTD.",
+    "license": "Apache-2.0",
+    "licenses": [
+        {
+            "type": "Apache-2.0",
+            "url": "http://www.apache.org/licenses/LICENSE-2.0"
+        }
+    ],
+    "homepage": "https://docs.ton.dev/",
+    "engines": {
+        "node": ">=6"
+    },
+    "dependencies": {
+        "@tonclient/appkit": "^0",
+        "@tonclient/core": "^1",
+        "@tonclient/lib-node": "^1",
+        "chalk": "^2.4.2",
+        "dockerode": "^3.3.1",
+        "fs-extra": "^9.1.0",
+        "mv": "^2.1.1",
+        "request": "^2.88.2",
+        "unzip-stream": "^0.3.1"
+    },
+    "devDependencies": {
+        "@babel/preset-env": "^7.14.7",
+        "@babel/preset-typescript": "^7.14.5",
+        "@types/dockerode": "^3.2.2",
+        "@types/fs-extra": "^9.0.8",
+        "@types/jest": "^27.4.1",
+        "@types/mv": "^2.1.1",
+        "@types/node": "^16.0.1",
+        "@types/request": "^2.48.5",
+        "@types/unzip-stream": "^0.3.0",
+        "@typescript-eslint/eslint-plugin": "^5.19.0",
+        "@typescript-eslint/parser": "^5.19.0",
+        "caxa": "^1.0.0",
+        "eslint": "^7.30.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-jest": "^26.1.4",
+        "eslint-plugin-prettier": "^4.0.0",
+        "husky": "^4.3.8",
+        "jest": "^27.5.1",
+        "lint-staged": "^12.3.8",
+        "prettier": "^2.3.2",
+        "rimraf": "^3.0.2",
+        "ts-jest": "^27.1.4",
+        "typescript": "^4.6.3"
+    }
 }
