@@ -1,11 +1,13 @@
 module.exports = {
-    "roots": [
-        "<rootDir>/src",
-    ],
-    "testMatch": [
-        "**/__tests__/**/*.+(ts|tsx|js)",
-    ],
-    "transform": {
-        "^.+\\.(ts|tsx)$": "ts-jest",
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.json',
     },
-};
+  },
+  moduleFileExtensions: ['ts', 'js'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testMatch: ['**/test/**/*.+(ts|js)', '**/__tests__/**/*.+(ts|tsx|js)'],
+  testEnvironment: 'node',
+}
