@@ -3,7 +3,7 @@ const header = `
  * This file was generated by EverDev.
  * EverDev is a part of EVER OS (see https://everos.dev).
  */
-`;
+`
 export const BasicContractCode =
     header +
     `
@@ -51,7 +51,7 @@ DEFINE_JSON_ABI(IProxy, DProxy, EProxy);
 
 // ----------------------------- Main entry functions ---------------------- //
 DEFAULT_MAIN_ENTRY_FUNCTIONS(Proxy, IProxy, DProxy, TIMESTAMP_DELAY)
-`;
+`
 export const BasicContractHeaders =
     header +
     `
@@ -81,4 +81,4 @@ __interface EProxy {
 };
 
 }} // namespace tvm::schema
-`;
+`
