@@ -1,4 +1,4 @@
-import { Terminal } from "../../core";
+import { Terminal } from "../../core"
 import {
     AbiFunction,
     AbiParam,
@@ -6,10 +6,10 @@ import {
     DecodedOutput,
     MessageBodyType,
     Signer,
-} from "@tonclient/core";
-import { Account } from "@tonclient/appkit";
-import { ParamParser } from "./param-parser";
-import { SignerRegistry } from "../signer/registry";
+} from "@tonclient/core"
+import { Account } from "@tonclient/appkit"
+import { ParamParser } from "./param-parser"
+import { SignerRegistry } from "../signer/registry"
 
 export async function resolveFunction(
     terminal: Terminal,
@@ -17,41 +17,46 @@ export async function resolveFunction(
     functionName: string,
     preventUi: boolean,
 ): Promise<AbiFunction> {
-    const functions = account.contract.abi.functions ?? [];
-    functionName = functionName.trim();
+    const functions = account.contract.abi.functions ?? []
+    functionName = functionName.trim()
     while (functionName === "" && !preventUi) {
-        terminal.log("\nAvailable functions:\n");
-        functions.forEach((x, i) => terminal.log(`  ${i + 1}) ${x.name}`));
-        terminal.log();
+        terminal.log("\nAvailable functions:\n")
+        functions.forEach((x, i) => terminal.log(`  ${i + 1}) ${x.name}`))
+        terminal.log()
 
-        const functionIndex = Number.parseInt(await inputParam(terminal, {
-            name: "Select function",
-            type: "number",
-        })) - 1;
+        const functionIndex =
+            Number.parseInt(
+                await inputParam(terminal, {
+                    name: "Select function",
+                    type: "number",
+                }),
+            ) - 1
         if (functionIndex >= 0 && functionIndex < functions.length) {
-            return functions[functionIndex];
+            return functions[functionIndex]
         }
-        terminal.log("Invalid function number. Try again.");
+        terminal.log("Invalid function number. Try again.")
     }
     if (functionName === "") {
-        throw new Error("Function name isn't specified");
+        throw new Error("Function name isn't specified")
     }
-    const func = account.contract.abi.functions?.find(x => x.name === functionName);
+    const func = account.contract.abi.functions?.find(
+        x => x.name === functionName,
+    )
     if (!func) {
-        throw new Error(`Function not found: ${functionName}`);
+        throw new Error(`Function not found: ${functionName}`)
     }
-    return func;
+    return func
 }
 
 function inputLine(terminal: Terminal, prompt: string): Promise<string> {
-    terminal.write(`  ${prompt}: `);
-    return new Promise((resolve) => {
-        const standard_input = process.stdin;
-        standard_input.setEncoding("utf-8");
+    terminal.write(`  ${prompt}: `)
+    return new Promise(resolve => {
+        const standard_input = process.stdin
+        standard_input.setEncoding("utf-8")
         standard_input.once("data", function (data) {
-            resolve(`${data}`.trim());
-        });
-    });
+            resolve(`${data}`.trim())
+        })
+    })
 }
 
 async function inputTuple(terminal: Terminal, param: AbiParam): Promise<any> {
@@ -67,36 +72,38 @@ async function inputTuple(terminal: Terminal, param: AbiParam): Promise<any> {
 
 async function inputScalar(terminal: Terminal, param: AbiParam): Promise<any> {
     while (true) {
-        const value = await inputLine(terminal, `${param.name} (${param.type})`);
+        const value = await inputLine(terminal, `${param.name} (${param.type})`)
         try {
-            return ParamParser.scalar(param, `"${value}"`);
+            return ParamParser.scalar(param, `"${value}"`)
         } catch (err: any) {
-            terminal.log(err.toString());
+            terminal.log(err.toString())
         }
     }
 }
 
 async function inputArray(terminal: Terminal, param: AbiParam): Promise<any[]> {
-    const item = JSON.parse(JSON.stringify(param)) as AbiParam;
-    item.type = param.type.slice(0, -2);
-    let count = Number(await inputLine(terminal, `Enter number of items in ${param.name}`));
-    const items = [];
-    let i = 1;
+    const item = JSON.parse(JSON.stringify(param)) as AbiParam
+    item.type = param.type.slice(0, -2)
+    let count = Number(
+        await inputLine(terminal, `Enter number of items in ${param.name}`),
+    )
+    const items = []
+    let i = 1
     while (i <= count) {
-        item.name = `${param.name} ${i}`;
-        items.push(await inputParam(terminal, item));
-        i += 1;
+        item.name = `${param.name} ${i}`
+        items.push(await inputParam(terminal, item))
+        i += 1
     }
-    return items;
+    return items
 }
 
 async function inputParam(terminal: Terminal, param: AbiParam): Promise<any> {
     if (param.type.endsWith("[]")) {
-        return inputArray(terminal, param);
-    } else if ( param.type.endsWith("tuple")) {
-        return inputTuple(terminal, param);
+        return inputArray(terminal, param)
+    } else if (param.type.endsWith("tuple")) {
+        return inputTuple(terminal, param)
     } else {
-        return inputScalar(terminal, param);
+        return inputScalar(terminal, param)
     }
 }
 
@@ -108,76 +115,93 @@ export async function resolveParams(
     preventUi: boolean,
 ): Promise<object> {
     if (paramsString.match(/{.+}/)) {
-        let jsonArgs: any;
+        let jsonArgs: any
         try {
-            jsonArgs = JSON.parse(paramsString);
+            jsonArgs = JSON.parse(paramsString)
         } catch (err: any) {
-            throw new Error(`Malformed JSON object has been passed`);
+            throw new Error(`Malformed JSON object has been passed`)
         }
-        terminal.log(`Skip ABI validation step because a JSON object has been passed as an argument.`);
+        terminal.log(
+            `Skip ABI validation step because a JSON object has been passed as an argument.`,
+        )
         return jsonArgs
     }
-    const values: { [name: string]: any } = ParamParser.components({
-        name: "params",
-        type: "tuple",
-        components: params,
-    }, paramsString);
-    let hasUserInput = false;
+    const values: { [name: string]: any } = ParamParser.components(
+        {
+            name: "params",
+            type: "tuple",
+            components: params,
+        },
+        paramsString,
+    )
+    let hasUserInput = false
     if (params.length > 0) {
-        terminal.log(prompt);
+        terminal.log(prompt)
     }
     for (const param of params) {
         if (param.name in values) {
-            terminal.log(`  ${param.name} (${param.type}): ${JSON.stringify(values[param.name])}`);
+            terminal.log(
+                `  ${param.name} (${param.type}): ${JSON.stringify(
+                    values[param.name],
+                )}`,
+            )
         }
     }
     for (const param of params) {
         if (!(param.name in values)) {
             if (!hasUserInput) {
                 if (preventUi) {
-                    throw new Error(`Missing parameter "${param.name}".`);
+                    throw new Error(`Missing parameter "${param.name}".`)
                 }
-                hasUserInput = true;
+                hasUserInput = true
             }
-            values[param.name] = await inputParam(terminal, param);
+            values[param.name] = await inputParam(terminal, param)
         }
     }
-    return values;
+    return values
 }
 
 export async function getRunParams(
     terminal: Terminal,
     account: Account,
     args: {
-        function: string,
-        input: string,
-        preventUi: boolean,
-        signer: string,
-        address: string,
-        runSigner: string,
+        function: string
+        input: string
+        preventUi: boolean
+        signer: string
+        address: string
+        runSigner: string
     },
 ): Promise<{
-    functionName: string,
-    functionInput: object,
-    signer: Signer,
+    functionName: string
+    functionInput: object
+    signer: Signer
 }> {
-    const func = await resolveFunction(terminal, account, args.function, args.preventUi);
+    const func = await resolveFunction(
+        terminal,
+        account,
+        args.function,
+        args.preventUi,
+    )
     const functionInput = await resolveParams(
         terminal,
         `\nParameters of ${func.name}:\n`,
         func.inputs,
         args.input,
         args.preventUi,
-    );
-    const signers = new SignerRegistry();
-    const signer = args.runSigner.trim() !== ""
-        ? await signers.resolveSigner(args.runSigner, { useNoneForEmptyName: false })
-        : account.signer;
+    )
+    const signers = new SignerRegistry()
+    const signer =
+        args.runSigner.trim() !== ""
+            ? await signers.resolveSigner(args.runSigner, {
+                  useNoneForEmptyName: false,
+              })
+            : account.signer
     return {
         functionName: func.name,
         functionInput,
         signer,
-    };
+    }
 }
 
 export async function logRunResult(
@@ -185,13 +209,16 @@ export async function logRunResult(
     decoded: DecodedOutput | undefined,
     transaction: any,
 ): Promise<void> {
-    const outMessages: DecodedMessageBody[] = (decoded?.out_messages as any) ?? [];
+    const outMessages: DecodedMessageBody[] =
+        (decoded?.out_messages as any) ?? []
     const details = {
         transaction,
         output: decoded?.output,
-        out_messages: outMessages.filter(x => x?.body_type !== MessageBodyType.Output),
-    };
-    terminal.log();
-    terminal.log('Execution has finished with result:');
-    terminal.log(JSON.stringify(details, undefined, "    "));
+        out_messages: outMessages.filter(
+            x => x?.body_type !== MessageBodyType.Output,
+        ),
+    }
+    terminal.log()
+    terminal.log("Execution has finished with result:")
+    terminal.log(JSON.stringify(details, undefined, "    "))
 }
