@@ -4,29 +4,25 @@ import {
     CommandArgVariant,
     Terminal,
     ToolController,
-} from "../../core";
-import {
-    SignerRegistry,
-} from "./registry";
-import {
-    TonClient,
-} from "@tonclient/core";
-import { formatTable } from "../../core/utils";
-import { NetworkRegistry } from "../network/registry";
+} from "../../core"
+import { SignerRegistry } from "./registry"
+import { TonClient } from "@tonclient/core"
+import { formatTable } from "../../core/utils"
+import { NetworkRegistry } from "../network/registry"
 
 const nameArg: CommandArg = {
     isArg: true,
     name: "name",
     title: "Signer name",
     type: "string",
-};
+}
 
 const secretArg: CommandArg = {
     isArg: true,
     name: "secret",
     title: "Secret key or seed phrase",
     type: "string",
-};
+}
 
 const mnemonicOpt: CommandArg = {
     name: "mnemonic",
@@ -34,7 +30,7 @@ const mnemonicOpt: CommandArg = {
     title: "Use mnemonic phrase",
     type: "boolean",
     defaultValue: "false",
-};
+}
 
 const dictionaryOpt: CommandArg = {
     name: "dictionary",
@@ -80,9 +76,9 @@ const dictionaryOpt: CommandArg = {
                 value: "8",
                 description: "Spanish",
             },
-        ];
+        ]
     },
-};
+}
 
 const wordsOpt: CommandArg = {
     name: "words",
@@ -90,7 +86,7 @@ const wordsOpt: CommandArg = {
     title: "Number of mnemonic words",
     type: "string",
     defaultValue: "12",
-};
+}
 
 const forceOpt: CommandArg = {
     name: "force",
@@ -98,63 +94,68 @@ const forceOpt: CommandArg = {
     title: "Overwrite signer if already exists",
     type: "boolean",
     defaultValue: "false",
-};
+}
 
 export const signerGenerateCommand: Command = {
     name: "generate",
     alias: "g",
     title: "Add signer with randomly generated keys",
-    args: [
-        nameArg,
-        mnemonicOpt,
-        dictionaryOpt,
-        wordsOpt,
-        forceOpt,
-    ],
-    async run(_terminal: Terminal, args: {
-        name: string,
-        mnemonic: boolean,
-        dictionary: string,
-        words: string,
-        force: boolean
-    }) {
+    args: [nameArg, mnemonicOpt, dictionaryOpt, wordsOpt, forceOpt],
+    async run(
+        _terminal: Terminal,
+        args: {
+            name: string
+            mnemonic: boolean
+            dictionary: string
+            words: string
+            force: boolean
+        },
+    ) {
         if (args.mnemonic) {
-            const dictionary = Number.parseInt(args.dictionary);
-            const word_count = Number.parseInt(args.words);
-            const phrase = (await TonClient.default.crypto.mnemonic_from_random({
+            const dictionary = Number.parseInt(args.dictionary)
+            const word_count = Number.parseInt(args.words)
+            const phrase = (
+                await TonClient.default.crypto.mnemonic_from_random({
+                    dictionary,
+                    word_count,
+                })
+            ).phrase
+            await new SignerRegistry().addMnemonicKey(
+                args.name,
+                "",
+                phrase,
                 dictionary,
-                word_count,
-            })).phrase;
-            await new SignerRegistry().addMnemonicKey(args.name, "", phrase, dictionary, args.force);
+                args.force,
+            )
         } else {
             await new SignerRegistry().addSecretKey(
                 args.name,
                 "",
-                (await TonClient.default.crypto.generate_random_sign_keys()).secret,
+                (
+                    await TonClient.default.crypto.generate_random_sign_keys()
+                ).secret,
                 args.force,
-            );
+            )
         }
     },
-};
+}
 
 export const signerAddCommand: Command = {
     name: "add",
     title: "Add signer",
-    args: [
-        nameArg,
-        secretArg,
-        dictionaryOpt,
-        forceOpt,
-    ],
-    async run(terminal: Terminal, args: {
-        name: string,
-        secret: string,
-        dictionary: string,
-        force: boolean
-    }) {
-        await new SignerRegistry().add(terminal, args);
+    args: [nameArg, secretArg, dictionaryOpt, forceOpt],
+    async run(
+        terminal: Terminal,
+        args: {
+            name: string
+            secret: string
+            dictionary: string
+            force: boolean
+        },
+    ) {
+        await new SignerRegistry().add(terminal, args)
     },
-};
+}
 
 export const signerListCommand: Command = {
     name: "list",
@@ -162,24 +163,24 @@ export const signerListCommand: Command = {
     title: "Prints list of registered signers",
     args: [],
     async run(terminal: Terminal, _args: {}) {
-        const registry = new SignerRegistry();
-        const networks = new NetworkRegistry();
-        const rows = [["Signer", "Public Key", "Used", "Description"]];
+        const registry = new SignerRegistry()
+        const networks = new NetworkRegistry()
+        const rows = [["Signer", "Public Key", "Used", "Description"]]
         registry.items.forEach(x => {
-            const summary = registry.getSignerSummary(x, networks);
+            const summary = registry.getSignerSummary(x, networks)
             rows.push([
                 summary.name,
                 summary.public,
                 summary.used,
                 summary.description,
-            ]);
-        });
-        const table = formatTable(rows, { headerSeparator: true });
+            ])
+        })
+        const table = formatTable(rows, { headerSeparator: true })
         if (table.trim() !== "") {
-            terminal.log(table);
+            terminal.log(table)
         }
     },
-};
+}
 
 export const signerGetCommand: Command = {
     name: "info",
@@ -193,22 +194,22 @@ export const signerGetCommand: Command = {
     ],
     async run(terminal: Terminal, args: { name: string }) {
         if (args.name === "") {
-            await signerListCommand.run(terminal, {});
-            return;
+            await signerListCommand.run(terminal, {})
+            return
         }
-        const signer = new SignerRegistry().get(args.name);
-        terminal.log(JSON.stringify(signer, undefined, "    "));
+        const signer = new SignerRegistry().get(args.name)
+        terminal.log(JSON.stringify(signer, undefined, "    "))
     },
-};
+}
 
 export const signerDeleteCommand: Command = {
     name: "delete",
     title: "Delete signer from registry",
     args: [nameArg],
     async run(_terminal: Terminal, args: { name: string }) {
-        new SignerRegistry().delete(args.name);
+        new SignerRegistry().delete(args.name)
     },
-};
+}
 
 export const signerDefaultCommand: Command = {
     name: "default",
@@ -216,9 +217,9 @@ export const signerDefaultCommand: Command = {
     title: "Set default signer",
     args: [nameArg],
     async run(_terminal: Terminal, args: { name: string }) {
-        new SignerRegistry().setDefault(args.name);
+        new SignerRegistry().setDefault(args.name)
     },
-};
+}
 
 export const SignerTool: ToolController = {
     name: "signer",
@@ -232,4 +233,4 @@ export const SignerTool: ToolController = {
         signerGetCommand,
         signerDefaultCommand,
     ],
-};
+}
