@@ -1,18 +1,15 @@
-import {
-    Component,
-} from "../../core";
+import { Component } from "../../core"
 
-const TOOL_FOLDER_NAME = "tonos-cli";
+const TOOL_FOLDER_NAME = "tonos-cli"
 
 export const components = {
-    tonoscli: new class extends Component {
+    tonoscli: new (class extends Component {
         getSourceName(version: string): string {
-            return `tonos-cli-${version.split(".").join("_")}-{p}.zip`;
+            return `tonos-cli-${version.split(".").join("_")}-{p}.zip`
         }
-    }(TOOL_FOLDER_NAME, "tonos-cli", {
+    })(TOOL_FOLDER_NAME, "tonos-cli", {
         resolveVersionRegExp: /tonos_cli\s+([0-9.]+)/,
         isExecutable: true,
         runGlobally: true,
     }),
-};
-
+}
