@@ -1,27 +1,39 @@
-import os from 'os'
-import { Component, Terminal } from '../../core'
-import { compareVersions, httpsGetJson, nullTerminal,  run } from '../../core/utils'
+import os from "os"
+import { Component, Terminal } from "../../core"
+import {
+    compareVersions,
+    httpsGetJson,
+    nullTerminal,
+    run,
+} from "../../core/utils"
 
-const TS4_PKG = 'tonos-ts4'
+const TS4_PKG = "tonos-ts4"
 const PYPI = `https://pypi.org/pypi/${TS4_PKG}/json`
 const currentOS = os.type()
-const [PYTHON, PIP] = ['Linux', 'Darwin'].includes(currentOS)
-    ? ['python3', 'pip3']
-    : ['python', 'pip']
+const [PYTHON, PIP] = ["Linux", "Darwin"].includes(currentOS)
+    ? ["python3", "pip3"]
+    : ["python", "pip"]
 
 export const components = {
-    ts4: new class extends Component {
+    ts4: new (class extends Component {
         async getCurrentVersion(): Promise<string> {
             let version
 
             try {
-                const output = await run(PIP, ['show', TS4_PKG], {}, nullTerminal)
-                version = output.split(os.EOL).find(line => (/^Version:/).exec(line))
+                const output = await run(
+                    PIP,
+                    ["show", TS4_PKG],
+                    {},
+                    nullTerminal,
+                )
+                version = output
+                    .split(os.EOL)
+                    .find(line => /^Version:/.exec(line))
             } catch {
                 // TODO handle the lack of 'pip'
                 console.debug(`Package ${TS4_PKG} not found`)
             }
-            return version ? version.split(/:\s*/)[1] : ''
+            return version ? version.split(/:\s*/)[1] : ""
         }
 
         async ensureVersion(
@@ -29,29 +41,36 @@ export const components = {
             force: boolean,
             requiredVersion?: string,
         ): Promise<boolean> {
-            const current = await this.getCurrentVersion();
+            const current = await this.getCurrentVersion()
             if (!force && current !== "" && !requiredVersion) {
-                return false;
+                return false
             }
-            let version = (requiredVersion ?? "latest").toLowerCase();
+            let version = (requiredVersion ?? "latest").toLowerCase()
             if (!force && version === current) {
-                return false;
+                return false
             }
-            const available = await this.loadAvailableVersions();
+            const available = await this.loadAvailableVersions()
             if (version === "latest") {
-                version = available[0];
+                version = available[0]
             } else {
                 if (!available.includes(version)) {
-                    throw new Error(`Invalid ${this.name} version ${version}`);
+                    throw new Error(`Invalid ${this.name} version ${version}`)
                 }
             }
             if (!force && version === current) {
-                return false;
+                return false
             }
-            const pkg = TS4_PKG + (version ? `==${version}` : '')
-            const output = await run(PIP, ['install', '-U', pkg], {}, nullTerminal)
+            const pkg = TS4_PKG + (version ? `==${version}` : "")
+            const output = await run(
+                PIP,
+                ["install", "-U", pkg],
+                {},
+                nullTerminal,
+            )
             const successPattern = `Successfully installed ${TS4_PKG}-${version}`
-            const isOk = output.split(os.EOL).find(line => line === successPattern)
+            const isOk = output
+                .split(os.EOL)
+                .find(line => line === successPattern)
 
             if (!isOk) {
                 terminal.writeError(output)
@@ -66,12 +85,12 @@ export const components = {
         async loadAvailableVersions(): Promise<string[]> {
             const info = await httpsGetJson(PYPI)
             const versions = Object.keys(info.releases)
-                .filter(v => (/^(\d+\.){2}\d+$/).test(v))
+                .filter(v => /^(\d+\.){2}\d+$/.test(v))
                 .sort(compareVersions)
                 .reverse()
             return versions
         }
-    }('', PYTHON, {
+    })("", PYTHON, {
         isExecutable: true,
         runGlobally: true,
     }),
