@@ -1,16 +1,13 @@
-import path from "path";
-import os from "os";
-import { TonClient } from "@tonclient/core";
-import { libNode } from "@tonclient/lib-node";
+import path from "path"
+import os from "os"
+import { TonClient } from "@tonclient/core"
+import { libNode } from "@tonclient/lib-node"
 
-export {
-    ComponentOptions,
-    Component,
-} from "./component";
+export { ComponentOptions, Component } from "./component"
 
 export type EverdevOptions = {
-    home?: string,
-};
+    home?: string
+}
 
 /**
  * Terminal object is implemented by `everdev` and passed to the controller's command handlers.
@@ -21,24 +18,24 @@ export interface Terminal {
      * Print line. Provided argument will be converted to strings and separated by space.
      * @param args values to print.
      */
-    log(...args: any[]): void,
+    log(...args: any[]): void
 
     /**
      * Prints text. No line feeds will be produced.
      * @param text Text that will be printed.
      */
-    write(text: string): void,
+    write(text: string): void
 
     /**
      * Prints error text. In case of CLI this text will be printed to stderr.
      * @param text Error text.
      */
-    writeError(text: string): void,
+    writeError(text: string): void
 }
 
 export type CommandArgVariant = {
-    value: string,
-    description?: string,
+    value: string
+    description?: string
 }
 
 /**
@@ -48,69 +45,69 @@ export type BaseCommandArg = {
     /**
      * Name of the command.
      */
-    name: string,
+    name: string
     /**
      * Short alias for an optional arg name.
      */
-    alias?: string,
+    alias?: string
     /**
      * Title of the argument.
      */
-    title?: string,
+    title?: string
     /**
      * Description of the argument.
      */
-    description?: string,
+    description?: string
     /**
      * Default value for the command. If missing then the user must specify value for this argument.
      */
-    defaultValue?: string,
+    defaultValue?: string
     /**
      * Determine that the arg is a CLI arg (not an option).
      */
-    isArg?: boolean,
+    isArg?: boolean
     /**
      * Get available CLI argument variants
      */
-    getVariants?(): CommandArgVariant[] | Promise<CommandArgVariant[]>,
-};
+    getVariants?(): CommandArgVariant[] | Promise<CommandArgVariant[]>
+}
 
 /**
  * Argument with path ti the file.
  */
 export type FileArg = BaseCommandArg & {
-    type: "file",
+    type: "file"
     /**
      * Determine files matched for this argument.
      */
-    nameRegExp: RegExp,
+    nameRegExp: RegExp
 }
 
 /**
  * Command argument with path to the folder.
  */
 export type FolderArg = BaseCommandArg & {
-    type: "folder",
+    type: "folder"
 }
 
 /**
  * String command argument.
  */
 export type StringArg = BaseCommandArg & {
-    type: "string",
+    type: "string"
 }
 
 /**
  * Boolean command argument.
  */
 export type BooleanArg = BaseCommandArg & {
-    type: "boolean",
+    type: "boolean"
 }
 
 /**
  * Command argument.
  */
-export type CommandArg = FileArg | FolderArg | StringArg | BooleanArg;
+export type CommandArg = FileArg | FolderArg | StringArg | BooleanArg
 
 /**
  * Command definition.
@@ -119,23 +116,23 @@ export interface Command {
     /**
      * Command name.
      */
-    name: string,
+    name: string
     /**
      * Command alias. Used in cli instead of full command name..
      */
-    alias?: string,
+    alias?: string
     /**
      * Command Title. Used in CLI short help and as a menu titles in IDE.
      */
-    title: string,
+    title: string
     /**
      * Description of the command.
      */
-    description?: string,
+    description?: string
     /**
      * Command argument definitions.
      */
-    args?: CommandArg[],
+    args?: CommandArg[]
 
     /**
      * Command handler.
@@ -143,22 +140,30 @@ export interface Command {
      *   Handler must print all human readable output using this terminal.
      * @param args Actual command arguments provided by user according to argument definitions.
      */
-    run(terminal: Terminal, args: any): Promise<void>,
+    run(terminal: Terminal, args: any): Promise<void>
 }
 
 type NameAlias = {
-    name: string,
-    alias?: string,
+    name: string
+    alias?: string
 }
 
-export function matchName(x: NameAlias, test: string | undefined | null,
+export function matchName(
+    x: NameAlias,
+    test: string | undefined | null,
 ): boolean {
-    test = (test || "").toLowerCase();
-    return x.name === test || x.alias === test;
+    test = (test || "").toLowerCase()
+    return x.name === test || x.alias === test
 }
 
-export function nameInfo(x: NameAlias, namePrefix = "", aliasPrefix = ""): string {
-    return x.alias ? `${namePrefix}${x.name}, ${aliasPrefix}${x.alias}` : `${namePrefix}${x.name}`;
+export function nameInfo(
+    x: NameAlias,
+    namePrefix = "",
+    aliasPrefix = "",
+): string {
+    return x.alias
+        ? `${namePrefix}${x.name}, ${aliasPrefix}${x.alias}`
+        : `${namePrefix}${x.name}`
 }
 
 /**
@@ -168,52 +173,54 @@ export interface ToolController {
     /**
      * Tool name. This is an identifier. Used only in configs.
      */
-    name: string,
+    name: string
     /**
      * Tool alias. Used in cli instead of full tool name..
      */
-    alias?: string,
+    alias?: string
     /**
      * Tool title. Human readable name of the tool.
      */
-    title?: string,
+    title?: string
     /**
      * Tool description.
      */
-    description?: string,
+    description?: string
     /**
      * Commands provided by tool controller.
      */
-    commands: Command[],
+    commands: Command[]
 }
 
 const config = {
-    home: path.resolve(os.homedir(), ".everdev")
-};
+    home: path.resolve(os.homedir(), ".everdev"),
+}
 
 export function everdevInit(options?: EverdevOptions) {
-    TonClient.useBinaryLibrary(libNode);
-    config.home = options?.home ?? config.home;
+    TonClient.useBinaryLibrary(libNode)
+    config.home = options?.home ?? config.home
 }
 
 export function everdevDone() {
-    TonClient.default.close();
+    TonClient.default.close()
 }
 
 /**
  * Home directory where tool must store all tool related resources.
  */
 export function everdevHome() {
-    return config.home;
+    return config.home
 }
 
-export async function getArgVariants(arg: CommandArg): Promise<CommandArgVariant[] | undefined> {
+export async function getArgVariants(
+    arg: CommandArg,
+): Promise<CommandArgVariant[] | undefined> {
     if (!arg.getVariants) {
-        return undefined;
+        return undefined
     }
-    const variants = arg.getVariants();
+    const variants = arg.getVariants()
     if (variants instanceof Promise) {
-        return await variants;
+        return await variants
     }
-    return variants;
+    return variants
 }
