@@ -1,3 +1,3 @@
-export {controllers, runCommand} from "./controllers/index";
+export { controllers, runCommand } from "./controllers/index"
 export * from "./core/utils"
 export * from "./core/index"
